Restructured the main page
